selectedEmotions updated to object. addLog functionality working.

+++ server/models/logEmotion.js
@@ -10,6 +10,6 @@ module.exports = {
             allowNull: false,
             primaryKey: true
         },
-        emotion_value: DataTypes.INTEGER
+        emotion_value: DataTypes.FLOAT
     })
 }
